reports_api/requirements.txt
fastapi
uvicorn[standard]
python-jose
clickhouse-driver
python-dotenv